I finally did the expenses and dashboard!!!

+++ app/Domain/Entity/Expense.php
@@ -9,11 +9,11 @@ use DateTimeImmutable;
 final class Expense
 {
     public function __construct(
-        public ?int $id,
-        public int $userId,
+        public ?int              $id,
+        public int               $userId,
         public DateTimeImmutable $date,
-        public string $category,
-        public int $amountCents,
-        public string $description,
+        public string            $category,
+        public float             $amountCents,
+        public string            $description,
     ) {}
 }
